+++ source/glannotations/include/glannotations/Utility/DistanceFieldGeneration.h
@@ -2,7 +2,7 @@
 
 #include <globjects/base/ref_ptr.h>
 
-#include <glannotations/PNGImage.h>
+#include <glannotations/Common/PNGImage.h>
 
 namespace glannotations {
 	namespace DistanceFieldGeneration {
@@ -13,7 +13,7 @@ namespace glannotations {
 
 		glannotations::PNGImage::colorVal_t colorValueFromFloat(double val);
 		globjects::ref_ptr<glannotations::PNGImage> bilinearResize(const glannotations::PNGImage& inImage, unsigned scaledWidth, unsigned scaledHeight);
-		globjects::ref_ptr<glannotations::PNGImage> bicubicResize(const glannotations::PNGImage& inImage, unsigned scaledWidth, unsigned scaledHeight);
+		globjects::ref_ptr<glannotations::PNGImage> bicubicResize(const glannotations::PNGImage& inImage, size_t scaledWidth, size_t scaledHeight);
 
 		namespace detail {
 			bool selfColoredNeighborsNot(const glannotations::PNGImage& original, unsigned x, unsigned y);
